+++ R/monitorLeaflet.R
@@ -39,13 +39,21 @@
 #' monitorLeaflet(CA_unhealthy_monitors, maptype="toner")
 #' }
 
-monitorLeaflet <- function(ws_monitor, slice=get('max'),
+monitorLeaflet <- function(ws_monitor,
+                           slice=get('max'),
                            breaks=AQI$breaks_24,
                            colors=AQI$colors,
                            labels=AQI$names,
                            legendTitle='Max AQI Level',
-                           radius=10, opacity=0.7, maptype="terrain",
+                           radius=10,
+                           opacity=0.7,
+                           maptype="terrain",
                            popupInfo=c('siteName','monitorID','elevation')) {
+  
+  # Sanity check
+  if ( monitor_isEmpty(ws_monitor) ) {
+    stop("ws_monitor object contains zero monitors")
+  }
   
   # BEGIN verbatim from monitor_map.R -----------------------------------------